# UpdateWeaponVectors MT