--- a/rts/Sim/Units/Unit.cpp
+++ b/rts/Sim/Units/Unit.cpp
@@ -668,8 +668,10 @@ void CUnit::Update()
 void CUnit::UpdateWeaponVectors()
 {
 	ZoneScoped;
+
 	if (!CanUpdateWeapons())
 		return;
+
 	for (CWeapon* w : weapons) {
 		w->UpdateWeaponVectors();
 	}
@@ -678,11 +680,10 @@ void CUnit::UpdateWeaponVectors()
 void CUnit::UpdateWeapons()
 {
 	ZoneScoped;
-	{
-		ZoneScopedN("CanUpdateWeapons");
-		if (!CanUpdateWeapons())
+
+	if (!CanUpdateWeapons())
 			return;
-	}
+
 	for (CWeapon* w: weapons) {
 		w->Update();
 	}