--- a/rts/Sim/Units/Unit.h
+++ b/rts/Sim/Units/Unit.h
@@ -195,6 +195,7 @@ public:
 	void UpdateLosStatus(int allyTeam);
 
 	void UpdateWeapons();
+	void UpdateWeaponVectors();
 
 	void SlowUpdateWeapons();
 	void SlowUpdateKamikaze(bool scanForTargets);