--- a/rts/Sim/Weapons/Weapon.cpp
+++ b/rts/Sim/Weapons/Weapon.cpp
@@ -254,14 +254,18 @@ void CWeapon::UpdateWeaponPieces(const bool updateAimFrom)
 }
 
 
-void CWeapon::UpdateWeaponVectors()
+void CWeapon::UpdateWeaponErrorVector()
 {
-	ZoneScoped;
-
+	// update conditional cause last SlowUpdate maybe longer away than UNIT_SLOWUPDATE_RATE
+	// i.e. when the unit got stunned (neither is SlowUpdate exactly called at UNIT_SLOWUPDATE_RATE, it's only called `close` to that)
 	float3 newErrorVector = (errorVector + errorVectorAdd);
 	if (newErrorVector.SqLength() <= 1.0f)
 		errorVector = newErrorVector;
+}
 
+void CWeapon::UpdateWeaponVectors()
+{
+	ZoneScoped;
 
 	relAimFromPos = owner->script->GetPiecePos(aimFromPiece);
 	owner->script->GetEmitDirPos(muzzlePiece, relWeaponMuzzlePos, weaponDir);
@@ -296,8 +300,6 @@ float CWeapon::GetPredictedImpactTime(float3 p) const
 void CWeapon::Update()
 {
 	ZoneScoped;
-	// update conditional cause last SlowUpdate maybe longer away than UNIT_SLOWUPDATE_RATE
-	// i.e. when the unit got stunned (neither is SlowUpdate exactly called at UNIT_SLOWUPDATE_RATE, it's only called `close` to that)
 
 	// Fast auto targeting needs to trigger an immediate retarget once the target is dead.
 	bool fastAutoRetargetRequired = fastAutoRetargeting && HaveTarget()