--- a/rts/Sim/Weapons/Weapon.h
+++ b/rts/Sim/Weapons/Weapon.h
@@ -96,6 +96,7 @@ public:
 	bool StopAttackingAllyTeam(const int ally);
 
 	bool IsFastAutoRetargetingEnabled() const { return fastAutoRetargeting; }
+	void UpdateWeaponVectors();
 
 protected:
 	virtual void FireImpl(const bool scriptCall) {}
@@ -107,7 +108,6 @@ protected:
 	static bool TargetInWater(const float3 tgtPos, const SWeaponTarget&);
 
 	void UpdateWeaponPieces(const bool updateAimFrom = true);
-	void UpdateWeaponVectors();
 	float3 GetLeadVec(const CUnit* unit) const;
 
 private: