This PR multithreads one of the most expensive parts of UpdateUnitWeapons, namely updating the weapon vectors. 
All units who have weapons perform this task every frame. The UpdateWeaponVectors function is self-contained, and does not call into lua or bos.

However, in order to make this optimization possible, the UpdateWeaponVectors function has to be called before fastautoretargeting and attacking existing targets. 
This is mostly due to the fact that Weapon::Update is virtual function, which would otherwise make any preMT/MT/postMT split near impossible. Also, most Weapon::Update overrides also do their own UpdateWeaponVectors, as does, for example Weapon::Attack. 

I have tested this change on the behaviour of a lot of units shooting, and found no unwanted side effects. 

Performance testing on two cases:
1. `/give 2000 armflash` These units are left idle. 

Old ST, 880 us per frame
![image](https://github.com/beyond-all-reason/spring/assets/109391/7d1bbee8-efe5-4c41-b7e7-8715beb4a532)

New MT, a total of 500 us per frame:
![image](https://github.com/beyond-all-reason/spring/assets/109391/d7a79937-8143-4c01-a3e1-1cb0a2325184)

2. `/luarules fightertest corak armpw` In this test, where almost all units are aiming and firing, 

Old ST: 1200 us per frame
New MT: 1050 us per frame.